fix(augmentation): RandomPlanckianJitter silently upcasts float16 input
Found via a systematic sweep of every Random* augmentation class for
dtype-preservation, run after fixing the analogous RandomMixUpV2 bug
(#4363) -- same symptom class, different mechanism.

apply_transform indexed the registered `pl` buffer (the Planckian-locus
color coefficient table) with `.to(device=input.device)` but no dtype=.
A registered buffer follows the whole nn.Module's `.to()` calls, but
stays at its own construction-time dtype (float32) if the caller never
explicitly `.to(dtype=...)`'s the augmentation module itself -- which
is the normal way these are used (call `aug(x)` directly on a float16
batch, never touching the module's own dtype). `input * r_w` then
silently upcasts a float16 input to float32 through ordinary PyTorch
type promotion the instant r_w/b_w (sliced from the buffer) aren't
already the same dtype -- reproduced directly.

Fixed by casting the buffer to input.dtype (not just input.device)
before indexing it.

## Context: found while auditing a systemic lead

While fixing #4363 (RandomMixUpV2) I flagged that the same
`_extract_device_dtype([self.<constructor-arg>])` pattern appears in
~20 other generator files, unverified. I ran a broader, more direct
sweep instead of auditing those files one at a time: constructed every
`kornia.augmentation.Random*` class and called it on a float16 batch,
checking output dtype. Out of 45 successfully-constructed-and-called
augmentations, exactly 2 leaked: this one, and `RandomDissolving`.
`RandomDissolving`'s leak comes from an internal Stable Diffusion
pipeline (a real, different, heavier case -- likely has genuine
numerical-stability considerations for running a diffusion UNet in
float16, not a simple tensor-dtype fix) -- not touched by this PR,
flagging it separately rather than rushing a fix there.

## Test plan

- [x] Added a test parametrized over float16/float32 confirming output
  dtype matches input dtype
- [x] Verified red-then-green
- [x] Full TestPlanckianJitter suite (6 tests) passes, ruff clean

+++ kornia/augmentation/_2d/intensity/planckian_jitter.py
@@ -193,8 +193,13 @@ class RandomPlanckianJitter(IntensityAugmentationBase2D):
     ) -> torch.Tensor:
         KORNIA_CHECK_SHAPE(input, ["*", "3", "H", "W"])
         # Index with the tensor itself: `.tolist()` reads the data, which graph capture cannot do. The buffer
-        # follows the module's device, so it is not re-assigned here.
-        coeffs = self.pl.to(device=input.device)[params["idx"].long()]
+        # follows the module's device, so it is not re-assigned here -- but NOT its dtype: `self.pl` is a
+        # registered buffer at its own construction-time dtype (float32), independent of whatever dtype a
+        # caller's `input` happens to be unless the whole nn.Module is explicitly `.to(dtype=...)`'d. Without
+        # an explicit dtype= here, `input * r_w` below silently upcasts a float16/bfloat16 input to float32
+        # the moment `r_w`/`b_w` (sliced from coeffs) aren't already the same dtype -- confirmed directly
+        # (RandomPlanckianJitter on a float16 batch returned a float32 batch).
+        coeffs = self.pl.to(device=input.device, dtype=input.dtype)[params["idx"].long()]
 
         r_w = coeffs[:, 0][..., None, None]
         b_w = coeffs[:, 1][..., None, None]
